divide write_sparameters_meep into 3 files, removed unnecessary files
