+++ gdsfactory/simulation/gmeep/__init__.py
@@ -2,36 +2,26 @@ import meep as mp
 
 from gdsfactory.config import logger
 from gdsfactory.simulation import plot
-from gdsfactory.simulation.gmeep.add_monitors import add_monitors
+from gdsfactory.simulation.get_sparameters_path import get_sparameters_data_meep
+from gdsfactory.simulation.gmeep import port_symmetries
 from gdsfactory.simulation.gmeep.get_simulation import get_simulation
-from gdsfactory.simulation.gmeep.get_transmission_2ports import (
-    get_transmission_2ports,
-    plot2D,
-    plot3D,
-)
-from gdsfactory.simulation.gmeep.plot_xsection import plot_xsection
-from gdsfactory.simulation.gmeep.write_sparameters_meep import (
-    write_sparameters_meep,
+from gdsfactory.simulation.gmeep.write_sparameters_meep import write_sparameters_meep
+from gdsfactory.simulation.gmeep.write_sparameters_meep_mpi import (
     write_sparameters_meep_mpi,
-    write_sparameters_meep_mpi_pool,
 )
-from gdsfactory.simulation.gmeep.write_sparameters_meep_1x1 import (
-    write_sparameters_meep_1x1,
+from gdsfactory.simulation.gmeep.write_sparameters_meep_mpi_pool import (
+    write_sparameters_meep_mpi_pool,
 )
 
 logger.info(f"Found Meep {mp.__version__!r} installed at {mp.__path__!r}")
 
 __all__ = [
-    "add_monitors",
     "get_simulation",
+    "get_sparameters_data_meep",
     "write_sparameters_meep",
     "write_sparameters_meep_mpi",
     "write_sparameters_meep_mpi_pool",
-    "write_sparameters_meep_1x1",
-    "get_transmission_2ports",
     "plot",
-    "plot2D",
-    "plot3D",
-    "plot_xsection",
+    "port_symmetries",
 ]
-__version__ = "0.0.2"
+__version__ = "0.0.3"
