+++ gdsfactory/simulation/gmeep/port_symmetries.py
@@ -1,0 +1,14 @@
+port_symmetries_1x1 = {
+    "o1": {
+        "s11": ["s22"],
+        "s21": ["s12"],
+    }
+}
+port_symmetries_crossing = {
+    "o1": {
+        "s11": ["s22", "s33", "s44"],
+        "s21": ["s12", "s34", "s43"],
+        "s31": ["s13", "s24", "s42"],
+        "s41": ["s14", "s23", "s32"],
+    }
+}
